+++ toshi_hazard_store/oq_import/aws_ecr_docker_image.py
