+++ toshi_hazard_store/oq_import/parse_oq_realizations.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING, Dict
 from nzshm_model import branch_registry
 from nzshm_model.psha_adapter.openquake import gmcm_branch_from_element_text
 
-from toshi_hazard_store.transform import parse_logic_tree_branches
+from .transform import parse_logic_tree_branches
 
 if TYPE_CHECKING:
     import pandas
